pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tfg-ml"
version = "0.1.0"
description = "Sistema multi‑agente para análisis y modelado de datos (TFG)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Autor TFG"}]
keywords = ["machine learning","multi-agent","crew","panel","tfg"]

dependencies = [
  "pandas>=2.0",
  "numpy>=1.26",
  "scikit-learn>=1.5",
  "pydantic>=2",
  "pydantic-settings>=2",
  "python-dotenv>=1",
  "joblib>=1.3",
]

[project.optional-dependencies]
ui = ["panel>=1.4"]
agents = ["crewai>=0.35.0","google-generativeai>=0.7.0"]
xgb = ["xgboost>=2.0"]
dev = ["pytest>=8","pytest-cov>=5","ruff>=0.5","black>=24.0","mypy>=1.10","pre-commit>=3.7"]

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
